+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# next [????.??.??]
+# 1.12 [2017.12.07]
 * Adapt to the `EmptyDataDeriving` proposal (introduced in GHC 8.4):
   * `Generics.Deriving.TH` now derives `to(1)` and `from(1)` implementations
     for empty data types that are strict in the argument.